trend_spotter/requirements.txt
google-adk
praw